First attempt at refactoring generic code into appropriate lib files
